Move extension logic into external file
